先这样
还差主界面，玩家生命判定，1p2p分数判定
道具系统（Buff）
新的敌机和道具

+++ Assets/Scripts/Tank/Spawner/TankSpawner.cs
@@ -90,7 +90,7 @@ public abstract class TankSpawner : MonoBehaviour
     /// 坦克阵亡时调用
     /// </summary>
     /// <param name="value"></param>
-    protected abstract void OnTankDie(float value);
+    protected abstract void OnTankDie(int value);
 
     /// <summary>
     /// 坦克生成时调用
